+++ sdk/monitor/azure-monitor-query/azure/monitor/query/_models.py
@@ -142,7 +142,7 @@ class MetricsResult(object):
             metrics=[Metric._from_generated(m) for m in generated.value] # pylint: disable=protected-access
         )
 
-class LogsQueryRequest(InternalLogQueryRequest):
+class LogsBatchQueryRequest(InternalLogQueryRequest):
     """A single request in a batch.
 
     Variables are only populated by the server, and will be ignored when sending a request.