+++ sdk/monitor/azure-monitor-query/tests/async/test_logs_client_async.py
@@ -2,7 +2,7 @@ import pytest
 import os
 from azure.identity.aio import ClientSecretCredential
 from azure.core.exceptions import HttpResponseError
-from azure.monitor.query import LogsQueryRequest
+from azure.monitor.query import LogsBatchQueryRequest
 from azure.monitor.query.aio import LogsQueryClient
 
 def _credential():
@@ -44,18 +44,18 @@ async def test_logs_batch_query():
     client = LogsQueryClient(_credential())
 
     requests = [
-        LogsQueryRequest(
+        LogsBatchQueryRequest(
             query="AzureActivity | summarize count()",
             timespan="PT1H",
             workspace_id= os.environ['LOG_WORKSPACE_ID']
         ),
-        LogsQueryRequest(
+        LogsBatchQueryRequest(
             query= """AppRequests | take 10  |
                 summarize avgRequestDuration=avg(DurationMs) by bin(TimeGenerated, 10m), _ResourceId""",
             timespan="PT1H",
             workspace_id= os.environ['LOG_WORKSPACE_ID']
         ),
-        LogsQueryRequest(
+        LogsBatchQueryRequest(
             query= "AppRequests | take 2",
             workspace_id= os.environ['LOG_WORKSPACE_ID']
         ),
@@ -90,19 +90,19 @@ async def test_logs_batch_query_additional_workspaces():
     query = "union * | where TimeGenerated > ago(100d) | project TenantId | summarize count() by TenantId"
 
     requests = [
-        LogsQueryRequest(
+        LogsBatchQueryRequest(
             query,
             timespan="PT1H",
             workspace_id= os.environ['LOG_WORKSPACE_ID'],
             additional_workspaces=[os.environ['SECONDARY_WORKSPACE_ID']]
         ),
-        LogsQueryRequest(
+        LogsBatchQueryRequest(
             query,
             timespan="PT1H",
             workspace_id= os.environ['LOG_WORKSPACE_ID'],
             additional_workspaces=[os.environ['SECONDARY_WORKSPACE_ID']]
         ),
-        LogsQueryRequest(
+        LogsBatchQueryRequest(
             query,
             workspace_id= os.environ['LOG_WORKSPACE_ID'],
             additional_workspaces=[os.environ['SECONDARY_WORKSPACE_ID']]